requirements.txt
numpy==1.26.4
pandas==2.2.0
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
pillow==10.2.0
ipython==8.18.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
scikit-image==0.22.0